h2spec add non-uring feature
