+++ crates/hring/Cargo.toml
@@ -38,3 +38,6 @@ tokio = { version = "1.24.2", default-features = false, features = ["io-util", "
 futures-util = { version = "0.3.25", default-features = false, features = ["std"] }
 curl = { version = "0.4.44", default-features = false, features = ["http2"] }
 libc = "0.2.139"
+
+[features]
+non-uring = []
