+++ test-crates/hring-h2spec/Cargo.toml
@@ -15,3 +15,7 @@ which = "4.4.0"
 
 [profile.dev.package."*"]
 opt-level = 2
+
+[features]
+# default = ["non-uring"]
+non-uring = ["hring/non-uring"]
